[fix] Fixes a bug in the "break" statement.
There was a bug in the "break" statement where the "OP_TEMP" instruction was not properly checked when ending a loop. "OP_TEMP" is afterall a simple integer, so it could be confused with an argument for any other instruction inside the loop body.
To solve this, if an instruction is found and it is not "OP_TEMP", we jump through its arguments as well.