[feat] Improves the allocation of a list literal.
Previously, the "DEF_LIST" instruction allocates a small list that was incremented as needed by the "PUSH_LIST" instruction. For a large, this would mean a huge number of instructions. Now, all elements are compiled and the list is allocated just once in "DEF_LIST" with the count of elements.

Additionally, the concatenation of lists using the "+" operator was removed for now.