[feat] Sets compiler/VM support for method inheritance.
Sets the compiler/VM support for method inheritance by using the "OP_INHERIT" instruction. It copies the methods from the superclass hashtable to the subclass. Moreover, it is emitted before the subclass content can be compiled. This way, methods from the subclass with the same name of the ones of the superclass will be automatically overrided.